+++ R/get_signature.R
@@ -10,11 +10,13 @@
 #' Function to obtain a SuperSig
 #' 
 #' Generate a tissue-specific SuperSig for a given dataset of mutations and
-#' exposure factor. Returns the SuperSig and a classification model trained with
-#' the SuperSig.
+#' exposure factor. Returns the SuperSig and a classification model trained 
+#' with the SuperSig.
 #' 
 #' @param dt a data frame of mutations (see vignette for details)
-#' @param factor the factor/exposure (e.g. "age", "smoking")
+#' @param factor the factor/exposure (e.g. "age", "smoking"). If the 
+#' factor = "age", the SuperSig is computed using
+#' counts. Otherwise, rates (counts/age) are used.
 #' @param wgs logical value indicating whether sequencing data is 
 #' whole-genome (wgs = \code{TRUE}) or whole-exome (wgs = \code{FALSE}) 
 #' 
@@ -29,27 +31,27 @@
 #' head(example_dt) # use example data from package
 #' input_dt <- make_matrix(example_dt) # convert to correct format
 #' input_dt$IndVar <- c(1, 1, 1, 0, 0) # add IndVar column
-#' get_signature(dt = input_dt, factor = "age") # get SuperSig
+#' get_signature(dt = input_dt, factor = "Age") # get SuperSig
 #' 
 get_signature <- function(dt, factor, wgs = FALSE){
     # Capitalize factor string
-    factor = toupper(factor)
+    factor <- toupper(factor)
     
     # Check column names of dt
     if(is.na(match("SAMPLE_ID",toupper(colnames(dt))))){
         stop('Input data frame missing sample_id column.')
     } else {
-        colnames(dt)[match("SAMPLE_ID",toupper(colnames(dt)))] = "sample_id"
+        colnames(dt)[match("SAMPLE_ID",toupper(colnames(dt)))] <- "sample_id"
     }
     if(is.na(match("AGE",toupper(colnames(dt))))){
         stop('Input data frame missing AGE column.')
     } else {
-        colnames(dt)[match("AGE",toupper(colnames(dt)))] = "AGE"
+        colnames(dt)[match("AGE",toupper(colnames(dt)))] <- "AGE"
     }
     if(is.na(match("INDVAR",toupper(colnames(dt))))){
         stop('Input data frame missing IndVar column.')
     } else {
-        colnames(dt)[match("INDVAR",toupper(colnames(dt)))] = "IndVar"
+        colnames(dt)[match("INDVAR",toupper(colnames(dt)))] <- "IndVar"
     }
     
     # Check number of samples (> 5 required)
@@ -64,21 +66,21 @@ get_signature <- function(dt, factor, wgs = FALSE){
     }
     
     # Get features
-    features_out = suppressWarnings(feature_selection(dt, factor, wgs))
+    features_out <- suppressWarnings(feature_selection(dt, factor, wgs))
     
     # Get apparent AUC and model
-    classification_out = supersig_classifier(dt = features_out$dt_new,
-                                             factor = factor,
-                                             keep_classifier = TRUE,
-                                             features_selected = 
-                                               features_out$features_selected,
-                                             select_n = features_out$select_n)
+    classification_out <- supersig_classifier(dt = features_out$dt_new,
+                                              factor = factor,
+                                              keep_classifier = TRUE,
+                                              features_selected = 
+                                                features_out$features_selected,
+                                              select_n = features_out$select_n)
     
     # Create and return S4 object
-    feature_names = names(classification_out$signature$mean_diffs)
-    out = SuperSig(Signature = classification_out$signature$mean_diffs,
-                   Features = features_out$new_partition[feature_names],
-                   AUC = classification_out$auc,
-                   Model = classification_out$classifier)
+    feature_names <- names(classification_out$signature$mean_diffs)
+    out <- SuperSig(Signature = classification_out$signature$mean_diffs,
+                    Features = features_out$new_partition[feature_names],
+                    AUC = classification_out$auc,
+                    Model = classification_out$classifier)
     return(out)
 }