+++ NEWS.md
@@ -1,5 +1,10 @@
 # supersigs 0.1.1
 
+* Add `partial_signature` function to remove signature contribution from data
+* Rename `dt` parameter as `data` in `get_signature`
+
+# supersigs 0.1.1
+
 * Add functionality to process VCF files with `VariantAnnotation::readVCF` and `process_vcf`
 * Change column names to ref/alt in `example_dt` and `make_matrix`
 * Add pre-trained SuperSigs as external data